Fix parsing issues form syntax nodes
Processing the syntax nodes is really hard. The new parsing uses the
semantic model more as it is much easier to process.

The commit fixes the semantic model issues up until now, by manually
adding the generated context source to the generator compilation.

It looks like the generator infrastructure excludes the generator
creatd files from the compilation. But we need our context base types
in the compiation.